fix: removes duplicate Page nodes appearing when adding a MenuItem
Previously duplicate Page objects would appear in the Content Object
selection list, which would occur after the queryset of objects is
filtered by the title. It was possible this would also affect other
content types, if the filter was across related objects. To avoid this,
adds a distinct() call to the queryset before it is returned if we have
a query to filter by. Adds tests which would previously have failed
without the change to add distinct().