feat: store expanded/collapsed for multiple menus
Updates the JS so that the ID of the menu is used when storing the
expanded/collapsed node ID's. This change means storing an array of IDs
for each menu as a json string. By storing session data unique for the
menu object it stops the session data being used for the wrong menu
object, and allows multiple menu states to be stored in a session.