Feature/expanded nodes are stored in the page session so that the structure of the nav tree is retained after editing a menu (#153)
* feat: store expanded/collapsed nodes in JS session object

When editing a menu, the expanded/collapsed nodes are stored in a
session object that lasts until the tab is closed. This allows the user
to edit the menu, or delete a node, and when the menu reloads, they
are displayed the menu with the same nodes still open.

* feat: store expanded/collapsed for multiple menus

Updates the JS so that the ID of the menu is used when storing the
expanded/collapsed node ID's. This change means storing an array of IDs
for each menu as a json string. By storing session data unique for the
menu object it stops the session data being used for the wrong menu
object, and allows multiple menu states to be stored in a session.

* fix: update changelog

Co-authored-by: Michael Collins <michaeljcollinsuk@users.noreply.github.com>